refactor(api): moved app file to /components
